# Calculate common tags only once.