This PR extracts commonTags outside of loop.
They are not changed inside the loop so there is no
need to recalculate them every time.
Ta reduce allocations tags slice i create with
initial capacity equal to the lenght of application labels.

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkApp_RegistrationIntents-8     438           415           -5.25%
BenchmarkApp_RegistrationIntents-8     436           419           -3.90%
BenchmarkApp_RegistrationIntents-8     432           415           -3.94%

benchmark                              old allocs     new allocs     delta
BenchmarkApp_RegistrationIntents-8     4              3              -25.00%
BenchmarkApp_RegistrationIntents-8     4              3              -25.00%
BenchmarkApp_RegistrationIntents-8     4              3              -25.00%

benchmark                              old bytes     new bytes     delta
BenchmarkApp_RegistrationIntents-8     128           208           +62.50%
BenchmarkApp_RegistrationIntents-8     128           208           +62.50%
BenchmarkApp_RegistrationIntents-8     128           208           +62.50%
```